apply tests: move helpers into common area
